Changed background color to web-supported color
The previous color used for the background, "web38", was unsupported on the web. As such, the text for the equation and its result were being lost, as the font was the same color as the default background. This has been fixed by changing the background color from "white38" to "green[900]" which is a very dark green.