Added 'http' package to dependency
In the commit that added remote calculation functionality (commit a40e9fd) a new package was needed to make HTTP requests. This needed to be reflected in the dependencies, which was a change that did not make it into that commit. This commit adds that new package to the dependency list.