Changed 'MyButton' to 'CalculatorButton' and moved class to different file.
The purpose behind this change was to figure out the logistics behind splitting the main application into multiple files. As the 'MyButton' class was an already-made self-contained class, it was the ideal candidate for this test. Turns out all that is needed is to add an import for the new file, and all of that file's classes will be accessible.

The naming change from "MyButton" to "CalculatorButton" was to make the name less generic, thereby making its purpose more intuitive.