Fixed split of 'CalculatorButton' from 'main.dart' into 'CalculatorButton.dart'
As per the title, the last commit attempted to take the 'CalculatorButton' class and put it in its own file. This was not successful, and that was due to the fact that both files were not in the same library. This change states that both `main.dart.` and `CalculatorButton.dart` are part of the `FlutterCalculator` library, fixing this issue.

Additonally, it appears that the last commit did not contain some select changes made to `CalculatorButton` that were meant to be a part of said commit. Those changes are all included in this commit.