DEL Button no longer removes last character of empty string
The method handling the DEL button on the calculator removed the last character of the input string regardless of content. This caused an error when it attempted to perform this operation on an empty string. While this would not crash the project, it would throw an exception at runtime regardless.

I decided to fix this as doing so was a simple task. This was because it only involved surrounding the culprit with an if-statement checking to see if the input string had a length greater than zero.