Added custom evaluator and fixed dividing
Made a quick-and-dirty expression evaluator for the project. Tested it with some PEMDAS and it seemed to work fine.

During this testing, I also found a bug when remotely evaluating an expression with division remotely. This was because the character for division, the forward slash, caused the URL to technically direct to a different page. Now, the frontend simply replaces the forward-slash with a "d" and the new custom evaluator treats "d" as the divisor function.